HEALING FINALLY DONE AHHHHH
You can do what ever the F#CK you want with the healing and the heal bar in the inspector

P/s: The heal bar might be misplaced due to the stupid canvas alignment ahh issue, don't mind that. Just click the "Heal Bar" in the "Player Canvas" in the "Player Character", and set its Y position from 0 to -185, thank you for reading.